requirements.txt
Cython>=0.29.1
dotmap>=1.2.20
flake8>=3.6.0
google-api-python-client>=1.6.5
h5py>=2.7.1
ipython>=6.2.1
jupyter_contrib_nbextensions>=0.4.0
jupyterlab>=0.31.12
keras==2.1.4
imgaug>=0.2.5
matplotlib>=2.2.0
msgpack==0.5.6
nbconvert>=5.3.1
nbformat>=4.4.0
nbstripout>=0.3.1
notebook>=5.4.0
numpy>=1.15.3
opencv-python>=3.4.0.12
pandas>=0.23.4
Pillow>=5.0.0
pycocotools>=2.0.0
pydicom>=1.0.2
scikit-learn>=0.20.0
scikit-image>=0.13.1
scipy>=1.0.0
tensorboard>=1.12.0
tensorflow-gpu>=1.12.0
tensorpack>=0.9.0.1
tqdm>=4.28.1
yapf>=0.24.0
